requirements.txt
djangorestframework
django-cors-headers
pyjwt
bcrypt